# openapi.yaml
...
openapi: 3.0.2
info:
  description:
    "This is a sample server Petstore server. For this sample, you can\
    \ use the api key `special-key` to test the authorization filters."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAPI Petstore
  version: 2.0.0
servers:
  - url: https://localhost:8090/v2
  - url: https://openapi.prod.localhost/api/v2
  # for tests validation middleware to work on random port.
  - url: /v2
...
paths:
...
  /pet:
    post:
      description: ""
      operationId: addPet
      requestBody:
        $ref: "#/components/requestBodies/Pet"
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Pet"
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
          description: successful operation
        "405":
          description: Invalid input
      security:
        - petstore_auth:
            - write:pets
            - read:pets
      summary: Add a new pet to the store
      tags:
        - pet
    put:
      description: ""
      operationId: updatePet
      requestBody:
        $ref: "#/components/requestBodies/Pet"
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Pet"
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
        "405":
          description: Validation exception
      security:
        - petstore_auth:
            - write:pets
            - read:pets
      summary: Update an existing pet
      tags:
        - pet
...
  /store/inventory:
    get:
      description: Returns a map of status codes to quantities
      operationId: getInventory
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  format: int32
                  type: integer
                type: object
          description: successful operation
      security:
        - api_key: []
      summary: Returns pet inventories by status
      tags:
        - store
  /store/order:
    post:
      description: ""
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
        description: order placed for purchasing the pet
        required: true
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Order"
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
          description: successful operation
        "400":
          description: Invalid Order
      summary: Place an order for a pet
      tags:
        - store
  /store/order/{orderId}:
    delete:
      description:
        For valid response try integer IDs with value < 1000. Anything
        above 1000 or nonintegers will generate API errors
      operationId: deleteOrder
      parameters:
        - description: ID of the order that needs to be deleted
          explode: false
          in: path
          name: orderId
          required: true
          schema:
            type: string
          style: simple
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Delete purchase order by ID
      tags:
        - store
    get:
      description:
        For valid response try integer IDs with value <= 5 or > 10. Other
        values will generated exceptions
      operationId: getOrderById
      parameters:
        - description: ID of pet that needs to be fetched
          explode: false
          in: path
          name: orderId
          required: true
          schema:
            format: int64
            maximum: 5
            minimum: 1
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/Order"
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
      summary: Find purchase order by ID
      tags:
        - store
  /user:
    post:
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              User:
                $ref: "#/components/examples/User"
        description: Created user object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResponse"
          description: successful operation
      # security:
      #   - api_key: []
      summary: creates a new user
      tags:
        - user
  /user/createWithArray:
    post:
      description: ""
      operationId: createUsersWithArrayInput
      requestBody:
        $ref: "#/components/requestBodies/UserArray"
      responses:
        default:
          description: successful operation
      security:
        - api_key: []
      summary: Creates list of users with given input array
      tags:
        - user
  /user/login:
    get:
      description: ""
      operationId: loginUser
      parameters:
        - description: The user name for login
          explode: true
          in: query
          name: username
          required: true
          schema:
            pattern: "^[a-zA-Z0-9]+[a-zA-Z0-9\\.\\-_]*[a-zA-Z0-9]+$"
            type: string
          style: form
        - description: The password for login in clear text
          explode: true
          in: query
          name: password
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          content:
            application/xml:
              schema:
                type: string
            application/json:
              schema:
                type: string
          description: successful operation
          headers:
            Set-Cookie:
              description:
                Cookie authentication key for use with the `api_key` apiKey
                authentication.
              explode: false
              schema:
                example: AUTH_KEY=abcde12345; Path=/; HttpOnly
                type: string
              style: simple
            X-Rate-Limit:
              description: calls per hour allowed by the user
              explode: false
              schema:
                format: int32
                type: integer
              style: simple
            X-Expires-After:
              description: date in UTC when token expires
              explode: false
              schema:
                format: date-time
                type: string
              style: simple
        "400":
          description: Invalid username/password supplied
      summary: Logs user into the system
      tags:
        - user
  /user/logout:
    get:
      description: ""
      operationId: logoutUser
      responses:
        default:
          description: successful operation
      security:
        - api_key: []
      summary: Logs out current logged in user session
      tags:
        - user
  /user/{username}:
    delete:
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - description: The name that needs to be deleted
          explode: false
          in: path
          name: username
          required: true
          schema:
            type: string
          style: simple
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      security:
        - api_key: []
      summary: Delete user
      tags:
        - user
    get:
      description: ""
      operationId: getUserByName
      parameters:
        - description: The name that needs to be fetched. Use user1 for testing.
          explode: false
          in: path
          name: username
          required: true
          schema:
            type: string
          style: simple
      responses:
        "200":
          content:
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          description: successful operation
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
      summary: Get user by user name
      tags:
        - user
    put:
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - description: name that need to be deleted
          explode: false
          in: path
          name: username
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        description: Updated user object
        required: true
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
      security:
        - api_key: []
      summary: Updated user
      tags:
        - user
...
components:
  requestBodies:
    UserArray:
      content:
        application/json:
          schema:
            items:
              $ref: "#/components/schemas/User"
            type: array
      description: List of user object
      required: true
    Pet:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Pet"
        application/xml:
          schema:
            $ref: "#/components/schemas/Pet"
        multipart/related:
          encoding:
            jsonData:
              contentType: application/json
              style: form
            binaryDataN2Information:
              contentType: application/vnd.3gpp.ngap
              headers:
                Content-Id:
                  explode: false
                  schema:
                    type: string
                  style: simple
              style: form
          schema:
            $ref: "#/components/schemas/updatePet_request"
      description: Pet object that needs to be added to the store
      required: true
  schemas:
...
    Order:
      description: An order for a pets from the pet store
      example:
        petId: 6
        quantity: 1
        id: 0
        shipDate: 2000-01-23T04:56:07.000+00:00
        complete: false
        status: placed
      properties:
        id:
          format: int64
          type: integer
        petId:
          format: int64
          type: integer
        quantity:
          format: int32
          type: integer
        shipDate:
          format: date-time
          type: string
        status:
          description: Order Status
          enum:
            - placed
            - approved
            - delivered
          type: string
        complete:
          default: false
          type: boolean
      title: Pet Order
      type: object
      xml:
        name: Order
    Category:
      description: A category for a pet
      x-go-custom-tag: acustomtag:"category"
      x-custom-comment: x-custom-comment:"test"
      example:
        name: name
        id: 6
      properties:
        id:
          format: int64
          type: integer
        name:
          pattern: "^[a-zA-Z0-9]+[a-zA-Z0-9\\.\\-_]*[a-zA-Z0-9]+$"
          type: string
      title: Pet category
      type: object
      xml:
        name: Category
    UpdateUserRequest:
      description: represents User data to update
      example:
        password: password
        email: email
        username: username
        firstName: John
        lastName: Doe
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        firstName:
          type: string
        lastName:
          type: string
      title: a User
      type: object
    CreateUserRequest:
      description: represents a new User
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          pattern: "^[ a-zA-Z0-9_-]+$"
          minLength: 3
        email:
          type: string
          pattern: "^[A-Za-z0-9+_.-]+@(.+)$"
        password:
          type: string
          minLength: 7
      title: a User
      type: object
    CreateUserResponse:
      description: represents the response to a User creation
      example:
        access_token: abcd
        user_id: 1
      properties:
        # go with jwt for now
        access_token:
          type: string
        user_id:
          format: int64
          type: integer
      type: object
    User:
      description: A User who is purchasing from the pet store
      properties:
        id:
          format: int64
          type: integer
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
        phone:
          type: string
        role:
          description: User role
          enum:
            - user
            - manager
            - admin
          type: string
      title: a User
      type: object
      xml:
        name: User
    Tag:
      description: A tag for a pet
      example:
        name: name
        id: 1
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
      title: Pet Tag
      type: object
      xml:
        name: Tag
    Pet:
      description: A pet for sale in the pet store
      example:
        photoUrls:
          - photoUrls
          - photoUrls
        name: doggie
        id: 0
        category:
          name: name
          id: 6
        tags:
          - name: name
            id: 1
          - name: name
            id: 1
        status: available
      properties:
        id:
          format: int64
          type: integer
          x-go-custom-tag: acustomtag:"id"
        category:
          $ref: "#/components/schemas/Category"
        name:
          example: doggie
          type: string
          x-go-custom-tag: acustomtag:"name"
        photoUrls:
          items:
            type: string
          type: array
          x-go-custom-tag: acustomtag:"photo_urls"
          xml:
            name: photoUrl
            wrapped: true
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          x-go-custom-tag: acustomtag:"tags"
          type: array
          xml:
            name: tag
            wrapped: true
        status:
          deprecated: true
          description: pet status in the store
          x-custom-comment: "This is a custom comment."
          x-go-custom-tag: acustomtag:"status"
          enum:
            - available
            - pending
            - sold
          type: string
      required:
        - name
        - photoUrls
      title: a Pet
      type: object
      xml:
        name: Pet
...
    updatePet_request:
      properties:
        jsonData:
          $ref: "#/components/schemas/Pet"
        binaryDataN2Information:
          format: binary
          type: string
      type: object
...
  examples:
    User:
      value:
        username: username
        email: email@email.com
        password: password
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://localhost:8090/api/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey